Seperate publisher into compiler components (#425)
* test-garden: add comment and svg

* refactor: move imports

* refactor: cleanup

* chore: fix tests

* chore: cleanup unused members
