+++ .prettierignore
@@ -15,3 +15,4 @@ package-lock.json
 package.json
 
 README.md
+*.md
