+++ src/dg-testVault/.obsidian/command-palette.json
@@ -1,0 +1,6 @@
+{
+  "pinned": [
+    "app:reload",
+    "digitalgarden:generate-garden-snapshot"
+  ]
+}
