+++ src/dg-testVault/.obsidian/workspace.json
@@ -28,12 +28,12 @@
             }
           },
           {
-            "id": "4384912f0352e092",
+            "id": "15581449e40a9e45",
             "type": "leaf",
             "state": {
               "type": "markdown",
               "state": {
-                "file": "001 Links.md",
+                "file": "E Embeds/E02 PNG.md",
                 "mode": "source",
                 "source": false
               }
@@ -62,29 +62,9 @@
                 "source": false
               }
             }
-          },
-          {
-            "id": "8898f2180fa76f65",
-            "type": "leaf",
-            "state": {
-              "type": "empty",
-              "state": {}
-            }
-          },
-          {
-            "id": "aaee9d4c117794c3",
-            "type": "leaf",
-            "state": {
-              "type": "markdown",
-              "state": {
-                "file": "000 Home.md",
-                "mode": "source",
-                "source": false
-              }
-            }
           }
         ],
-        "currentTab": 6
+        "currentTab": 3
       }
     ],
     "direction": "vertical"
@@ -150,7 +130,7 @@
             "state": {
               "type": "backlink",
               "state": {
-                "file": "000 Home.md",
+                "file": "002 Hidden page.md",
                 "collapseAll": false,
                 "extraContext": false,
                 "sortOrder": "alphabetical",
@@ -167,7 +147,7 @@
             "state": {
               "type": "outgoing-link",
               "state": {
-                "file": "000 Home.md",
+                "file": "002 Hidden page.md",
                 "linksCollapsed": false,
                 "unlinkedCollapsed": true
               }
@@ -190,7 +170,7 @@
             "state": {
               "type": "outline",
               "state": {
-                "file": "000 Home.md"
+                "file": "002 Hidden page.md"
               }
             }
           }
@@ -203,21 +183,30 @@
   },
   "left-ribbon": {
     "hiddenItems": {
+      "obsidian-excalidraw-plugin:Create new drawing": false,
       "digitalgarden:Digital Garden Publication Center": false,
       "switcher:Open quick switcher": false,
       "graph:Open graph view": false,
       "canvas:Create new canvas": false,
       "daily-notes:Open today's daily note": false,
       "templates:Insert template": false,
-      "command-palette:Open command palette": false,
-      "obsidian-excalidraw-plugin:Create new drawing": false
+      "command-palette:Open command palette": false
     }
   },
-  "active": "aaee9d4c117794c3",
+  "active": "16078f5ef00e92e4",
   "lastOpenFiles": [
-    "004 Publishing this garden.md",
-    "003 Non published page.md",
+    "E Embeds/E02 PNG.md",
     "000 Home.md",
+    "005 Custom filters.md",
+    "002 Hidden page.md",
+    "E Embeds/garden-gate.svg",
+    "E Embeds/E02 Image",
+    "E Embeds/ahven_logo.png",
+    "E Embeds/IMG_8AA032A19D75-1.jpeg",
+    "E Embeds/E01 SVG.md",
+    "E Embeds",
+    "009 Comments.md",
+    "001 Links.md",
     "Path Rewriting/Subfolder2/More specific path rewriting.md",
     "Path Rewriting/Subfolder/How deep do the rewrite rules go?.md",
     "Path Rewriting/Subfolder2",
@@ -236,12 +225,9 @@
     "006 Custom title.md",
     "F Folder/F01 Note in folder!.md",
     "F Folder/004 Folder set to root.md",
-    "F Folder",
-    "001 Links.md",
-    "002 Hidden page.md",
-    "Path Rewriting/F01 Note in folder!.md",
-    "005 Custom filters.md",
+    "004 Publishing this garden.md",
+    "003 Non published page.md",
     "Folder/004 Folder set to root.md",
-    "Folder"
+    "F Folder"
   ]
 }
