+++ src/dg-testVault/009 Comments.md
@@ -1,0 +1,6 @@
+---
+dg-publish: true
+---
+This is the only content on this page
+
+%% THEY'LL NEVER FIND ME!!  %%
