+++ src/dg-testVault/E Embeds/E01 SVG.md
@@ -1,0 +1,3 @@
+Here's an SVG embed: 
+
+![[garden-gate.svg]]
